After discussion with George, I implemented b2bua functionality for registered Restcomm clients. Before this patch clients had to always go through an app specified in the VoiceUrl parameter. The app then determined how client calls are handled. This caused a lot of confusion on the team, even though there are cases where this feature makes sense. For example allowing the application to control client calls based on certain business rules, e.g. don't allow the client to call after 5pm or make international calls.
With this patch, Restcomm continues to behave as it did when a VoiceURL is specified for a Client entry. However its not mandatory. When the VoiceUrl parameter is left empty, Restcomm allows registered Clients to make direct peer to peer calls, without routing media via Restcomm.

This not only allows the flexibility to use Restcomm as a signaling B2BUA for SIP, but it also enables WebRTC p2p with cutting edge advanced media codecs for voice and video that may not be supported yet by the server side media server that Restcomm is configured to work with.

The issue is further tracked here:
https://bitbucket.org/telestax/telscale-restcomm/issue/115/restcomm-mandates-voiceurl-for-all