Fix Usage endpoint (call duration) for  MariaDB
Implement changes to have the Usage endpoint and namely the duration
field for calls to work properly both at HSQLDB and MariaDB.

Needed to break the query where the column name goes as MariaDB won't
accept " in column name and HSQLDB requires it. In the proper usage.xml
it is being added or not, as required.

Fixes #309