Copy-pasted scripts from Restcomm-Docker project
Moving them into this GH repo, so they can be included in the docker image build.
As we are moving towards using the docker image as our binary (instead of just the WAR file), it makes sense to keep all those components that are used to build the image in the same repository anyway.