Split the maven build in 2 profiles `standalone` and `all-in-one`
In a nutshell, the `standalone` profile includes just restcomm-connect. The `all-in-one` includes RC + RMS + RVD + WebRTC UI, just like it used to be before introducing these 2 profiles. (Eventually, we should be getting rid of this second one, but for the moment, for backwards compatibility and for moving forward with a smoother migration process, we should keep it around and just focus on improving the WAR + docker image generation on that part (in maven build, instead of `ant` + separate restcomm-docker repo).