Moved docker image generation to `package` maven phase
the bindings to the `pre-integration-test` and `post-integration-test` phases only really make sense if/when we want to start some containers during the `verify` phase.
I am keeping those execution bindings in, so we have the hooks ready. However, you can *safely remove* the `start` and `stop` executions of the fabric8 docker maven plugin with no problems.